% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{box_simple}
\alias{box_simple}
\title{Boxplot}
\usage{
box_simple(fill = "deepskyblue4", ...)
}
\description{
This function creates a boxplot with pre-defined colors
}
\examples{
ggplot(mpg, aes(drv, hwy)) + box_simple()
}
