% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{histogram}
\alias{histogram}
\title{Histogram}
\usage{
histogram(fill = "deepskyblue4", ...)
}
\description{
This function creates a histogram with pre-defined colors
}
\examples{
ggplot(mpg, aes(hwy)) + histogram()
}
