% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{linear}
\alias{linear}
\title{Linear function for ScatterLinear}
\usage{
linear(color = "deepskyblue4", se = FALSE, ...)
}
\description{
This function adds a linear fit line to data
}
\examples{
ggplot(mpg, aes(displ, hwy)) + scatter_fit() + nonlinear()
}
