% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{nonlinear}
\alias{nonlinear}
\title{NonLinear function for ScatterLinear}
\usage{
nonlinear(color = "deepskyblue4", se = FALSE, ...)
}
\description{
This function adds a non-linear fit line to data
}
\examples{
ggplot(mpg, aes(displ, hwy)) + scatter_fit() + nonlinear()
}
