% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scatter}
\alias{scatter}
\title{Scatter function}
\usage{
scatter(color = "deepskyblue4", ...)
}
\arguments{
\item{color}{color of the scatter points}
}
\value{
scatter plot
}
\description{
This function creates a scatter plot with using "deepskyblue4" color
}
\examples{
ggplot(mpg, aes(displ, hwy)) + scatter()
}
