% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scatter_fit}
\alias{scatter_fit}
\title{define scatter function with linear fit}
\usage{
scatter_fit(...)
}
\description{
This function modifies a pre-defined theme (theme_bw), and make it a bit simple.
}
\examples{
ggplot(mpg, aes(displ, hwy)) + theme_simple()
}
