% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_simple}
\alias{theme_simple}
\title{Building a simpler ggplot theme}
\usage{
theme_simple()
}
\description{
This function modifies a pre-defined theme (theme_bw), and make it a bit simple.
}
\examples{
ggplot(mpg, aes(displ, hwy)) + theme_simple()
}
